Backend/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["thesloppyguy <sahilmandal@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.2"
djangorestframework = "^3.15.2"
markdown = "^3.7"
django-filter = "^24.3"
djangorestframework-api-key = "^3.0.0"
django-cors-headers = "^4.5.0"
django-redis = "^5.4.0"
psycopg = "^3.2.3"
pydantic = "^2.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
